+++ CHANGELOG.md
@@ -5,8 +5,13 @@
 ### Changed
 
 - [TD-6921] Download/Upload of fields of type link must include both name and url
+- [TD-6022] Upload of hierarchy fields should validate that the value is a valid one
 
 ## [7.0.0] 2025-01-13
+
+### Added
+
+- [TD-6022] Add hierarchy validation during business concept Excel load
 
 ### Added